+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # vNext
 
+* Fix: Remove method reference in SentryEnvelopeItem (#1091)
 * Enhancement: Set transaction name on events and transactions sent using Spring integration (#1067) 
 * Fix: Set current thread only if there are no exceptions
 * Enhancement: Set global tags on SentryOptions and load them from external configuration (#1066)
@@ -11,6 +12,7 @@
 * Enhancement: Make InAppIncludesResolver public (#1084)
 * Ref: Change "op" to "operation" in @SentrySpan and @SentryTransaction
 * Fix: SentryOptions creates GsonSerializer by default
+* Fix: Append DebugImage list if event already has it
 
 # 4.0.0-alpha.1